Implementeaza clasa URLNormalization
Issue: #22
Implementeaza metodele necesare normalizarii de URL
